moved test to the fix_env file
