+++ desktop/src-tauri/src/fix_env.rs
@@ -60,3 +60,19 @@ pub fn fix_env(var_name: &str) -> Result<(), Error> {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    #[test]
+    fn test_fix_env() {
+        std::env::set_var("_TEST_VARIABLE", "test_value");
+        assert_eq!(std::env::var("_TEST_VARIABLE").unwrap(), "test_value");
+    }
+
+    #[test]
+    fn test_fix_env_multiline() {
+        std::env::set_var("_TEST_VARIABLE", "test_value\nnew_line\nanother_line");
+        assert_eq!(std::env::var("_TEST_VARIABLE").unwrap(), "test_value\nnew_line\nanother_line");
+    }
+}
